refactor: moving wrapTimeout to its own file
