+++ src/index.ts
@@ -1,4 +1,3 @@
-export * from './abort'
 export * from './checkpoint'
 export * from './composeAbort'
 export * from './cleanupPromise'
@@ -9,3 +8,4 @@ export * from './isPromiseLike'
 export * from './raceWithSignal'
 export * from './toPromise'
 export * from './types'
+export * from './wrapTimeout'
